Ensure that Python does not cast integers into floats.
Former-commit-id: 1b73d9043ad278a9e3743b272cdd28e11ad95751

+++ python/peano4/toolbox/blockstructured/BackupPatchOverlap.py
@@ -30,7 +30,7 @@ class BackupPatchOverlap(ActionSet):
     self.d = {}
     self.d[ "UNKNOWNS" ]           = str(patch_overlap_in.no_of_unknowns)
     self.d[ "DOFS_PER_AXIS" ]      = str(patch_overlap_in.dim[1])
-    self.d[ "OVERLAP" ]            = str(patch_overlap_in.dim[0]/2)
+    self.d[ "OVERLAP" ]            = str(int(patch_overlap_in.dim[0]/2))
     self.d[ "FACES_ACCESSOR_IN" ]  = "fineGridFace"  + patch_overlap_in.name
     self.d[ "FACES_ACCESSOR_OUT" ] = "fineGridFace"  + patch_overlap_out.name
     self.d[ "GUARD_PREDICATE" ]    = guard
